requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
frozenlist==1.3.3
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
httpcore==0.13.2
httpx==0.18.1
hyperframe==5.2.0
idna==2.10
imageio==2.26.1
imageio-ffmpeg==0.4.8
MouseInfo==0.1.3
moviepy==1.0.3
multidict==6.0.4
natsort==8.3.1
numpy==1.24.2
openai==0.27.1
opencv-python==4.7.0.72
Pillow==9.4.0
proglog==0.1.10
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
requests==2.28.2
rfc3986==1.5.0
robloxapi==4.7
sniffio==1.3.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.14
yarl==1.8.2
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
frozenlist==1.3.3
h11==0.8.1
h2==3.2.0
hpack==3.0.0
http3==0.6.7
httpcore==0.13.2
httpx==0.18.1
hyperframe==5.2.0
idna==2.10
imageio==2.26.1
imageio-ffmpeg==0.4.8
MouseInfo==0.1.3
moviepy==1.0.3
multidict==6.0.4
natsort==8.3.1
numpy==1.24.2
openai==0.27.1
opencv-python==4.7.0.72
Pillow==9.4.0
proglog==0.1.10
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytweening==1.0.4
requests==2.28.2
rfc3986==1.5.0
robloxapi==4.7
sniffio==1.3.0
soupsieve==2.4
tqdm==4.65.0
urllib3==1.26.14
yarl==1.8.2
